# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "NoteMesh Api",
    "description": "Note handling API",
    "version": "1.0"
  },
  "basePath": "/api",
  "paths": {
    "/notes": {
      "get": {
        "summary": "Get all notes",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Note"
              }
            }
          }
        },
        "tags":["Notes"]
      },
      "post": {
        "summary": "Post the note",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Note"
              }
            }
          }
        },
        "tags":["Notes"]
      },
      "responses":{
        "200": {
          "description": "OK",
          "schema": {
            "type": "array",
            "items": {
            "$ref": "#/definitions/Note"
            }
          }
        }
      },
      "tags" : ["Notes"]
    },
    "/notes/{noteId}": {
      "get": {
        "summary": "Get a note by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the note to retrieve",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Note"
            }
          },
          "404": {
            "description": "Note not found"
          }
        },
      "tags":["Notes"]
      },
      "delete":{
        "summary":"Delete the note of that id",
                "responses": {
          "204": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Note"
            }
          },
          "404": {
            "description": "Note not found"
          }
        },
        "tags":["Notes"]
      }
    },
    "/user": {
      "POST":{
        "summary": "save the user",
        "responses":{
          "200":{
            "description" : "OK",
            "schema":{

            }
          },
          "404":{
            "description":"user not found"
          }
        },
        "tags":["User"]
      }
    }
  },
  "definitions": {
    "Note": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "content": {
          "type": "string"
        }
      }
    }
  },
  "tags": [
    {
      "name": "Notes",
      "description": "Operations related to notes"
    },
    {
      "name": "Users",
      "description": "Operations related to users"
    },
    {
      "name": "Authentication",
      "description": "Operations related to authentication and login"
    }
  ]
}